mark @InternalApi, move shared code to Util file,
